+++ jsonargparse/_deprecated.py
@@ -25,6 +25,7 @@ __all__ = [
     "LoggerProperty",
     "ParserError",
     "get_config_read_mode",
+    "namespace_to_dict",
     "null_logger",
     "set_docstring_parse_options",
     "set_config_read_mode",
@@ -689,3 +690,14 @@ class ActionJsonnetExtVars:
 )
 class LoggerProperty(InternalLoggerProperty):
     """Adds a logger property, intended for internal use."""
+
+
+@deprecated(
+    """
+    namespace_to_dict was deprecated in v4.40.0 and will be removed in v5.0.0.
+    Instead you can use ``.clone().as_dict()`` or ``.as_dict()``.
+"""
+)
+def namespace_to_dict(namespace: Namespace) -> Dict[str, Any]:
+    """Returns a copy of a nested namespace converted into a nested dictionary."""
+    return namespace.clone().as_dict()